requirements.txt
asgiref
beautifulsoup4==4.13.4
cbor==1.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
Django
elasticsearch==8.11.0
idna==3.10
ijson==3.4.0
inscriptis==2.6.0
ir_datasets==0.5.10
lxml==5.4.0
lz4==4.4.4
numpy==2.2.5
pyarrow==20.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
search_engines==1.0.7
setuptools==78.1.1
six==1.17.0
soupsieve==2.7
sqlparse 
tqdm==4.67.1
trec-car-tools==2.6
typing_extensions 
unlzw3==0.2.3
urllib3==2.4.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wheel==0.45.1
zlib-state==0.1.9
